requirements.txt
# Core dependencies
anthropic>=0.40.0
langchain-anthropic>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.2.0
openai>=1.50.0
spacy>=3.7.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Web scraping and content extraction
requests>=2.31.0
requests-cache>=1.1.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
feedparser>=6.0.0

# Language detection
langdetect>=1.0.9

# Utilities
python-dateutil>=2.8.0
